requirements.txt
Django==5.2.3
djangorestframework==3.16.0
psycopg2-binary==2.9.10
django-cors-headers==3.18.0
